% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{circleFun}
\alias{circleFun}
\title{circleFun

collecting data frame of arc/point based on the diameter/radius and angle according to x-y axis
\code{circleFun} circle function for collecting a dataset of dots for a circle}
\usage{
circleFun(center = c(0, 0), r = 10, npoints = 100)
}
\arguments{
\item{center}{center of the circle}

\item{r}{radius of the circle}

\item{npoints}{a sequence of number with a distance of "by" by = ((to - from)/(length.out - 1)),}
}
\value{
data frame of dots x and y coordinate for the circle
}
\description{
circleFun

collecting data frame of arc/point based on the diameter/radius and angle according to x-y axis
\code{circleFun} circle function for collecting a dataset of dots for a circle
}
\examples{

}
