% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedPlotting.R
\name{combinedPlotting}
\alias{combinedPlotting}
\title{The main plotting function to combine the two parts of the workflow}
\usage{
combinedPlotting(inputCSV, inputTree)
}
\arguments{
\item{inputCSV}{the generated csv file with present and absent of VF of each strain}

\item{inputNWK}{the generated nwk file of the relationship between strains}
}
\value{
Combined plotting of the data from csv and tree from newick file
}
\description{
\code{combinedPlotting} The function will generate a newick file with relationships between genome
}
\examples{

}
