% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{getCoordinates}
\alias{getCoordinates}
\title{getCoordinates

This function take input of the tree data and layer data and autput the tree
data with coordinate info of each node get Coordinates}
\usage{
getCoordinates(data1, data2, npoint)
}
\arguments{
\item{data1}{tree data frame}

\item{data2}{layer data frame contaions all layers data}

\item{npoint}{the number of points of each layer}
}
\value{
updated tree data frame with x and y coordinates
}
\description{
getCoordinates

This function take input of the tree data and layer data and autput the tree
data with coordinate info of each node get Coordinates
}
\examples{

}
