% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringPlot.R
\name{getRingRadius}
\alias{getRingRadius}
\title{getRingRadius}
\usage{
getRingRadius(ring_table_data, distance_between_tree_and_ring = 1,
  distance_between_ring_and_ring = 1, tree_outer_radius)
}
\arguments{
\item{ring_table_data}{data frame containing all information about the input table/csv}

\item{distance_between_tree_and_ring}{this is the unit distance/ radius difference between the tree and the first ring}

\item{distance_between_ring_and_ring}{this is the unit distance/ radius difference between rings}

\item{tree_outer_radius}{tree outer radius}
}
\value{
data frame containing responding radius and rings
}
\description{
getRingRadius
}
\examples{

}
