% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table100}
\alias{Table100}
\title{Table100}
\format{A data frame with 100 rows and 4 columns:
\describe{
\item{X}{id1}
\item{id}{id2}
\item{vf1}{virulence factor 1 potential of the strain with id1/id2}
\item{vf2}{virulence factor 2 potential of the strain with id1/id2}

...
}}
\usage{
Table100
}
\description{
A dataset contain 100 factor tree node info
}
\keyword{datasets}
