% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table150}
\alias{Table150}
\title{Table150}
\format{A data frame with 150 rows and 4 columns:
\describe{
\item{X}{id1}
\item{id}{id2}
\item{vf1}{virulence factor 1 potential of the strain with id1/id2}
\item{vf2}{virulence factor 2 potential of the strain with id1/id2}

...
}}
\usage{
Table150
}
\description{
A dataset contain 150 factor tree node info
}
\keyword{datasets}
