% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table25}
\alias{Table25}
\title{Table25}
\format{A data frame with 25 rows and 4 columns:
\describe{
\item{X}{id1}
\item{id}{id2}
\item{id}{id2}

...
}}
\usage{
Table25
}
\description{
A dataset contain 25 factor tree node info
}
\keyword{datasets}
