% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringPlot.R
\name{tableInputProcess}
\alias{tableInputProcess}
\title{tableInputProcess}
\usage{
tableInputProcess(input_name)
}
\arguments{
\item{inputname}{the string of the location of the input table/csv}
}
\value{
data frame containing all information about the input table/csv
}
\description{
\code{tree_input_process} preprocess the phylo object from the input tree file with newick format
}
\examples{

}
