% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tree150}
\alias{Tree150}
\title{Tree150}
\format{A phylo object with 4 main parts
\describe{
\item{edge}{The matrix edge contains the beginning and ending node number for all the nodes and tips in the tree. By convention, the tips of the tree are numbered 1 through n for n tips; and the nodes are numbered n + 1 through n + m for m nodes. m = n - 1 for a fully bifurcating tree. This is just to keep track of which nodes are internal and which are leaves.}
\item{tip.label}{The vector tip.label contains the labels for all the tips in the tree. The order of tip.label is the order of the tips numbered 1 through n in edge.}
\item{edge.length}{a vector of class "numeric" containing all the edge lengths of the tree in the same order as the rows in edge; and root.edge, a numeric value giving the length of the root edge, if one exists. }
\item{Nnode}{The integer Nnode contains the number of internal nodes in the tree, including the root of the tree if the tree is rooted.}

...
}}
\source{
http://www.phytools.org/eqg/Exercise_3.2/
}
\usage{
Tree150
}
\description{
A "phylo" object contain information about the tree
created from the sample150.newick file contains taxonomy info of all strains
}
\keyword{datasets}
