% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{tree_input_process}
\alias{tree_input_process}
\title{tree_input_process

\code{tree_input_process} preprocess the phylo object from the input tree file with newick format}
\usage{
tree_input_process(inputname = "sample1.newick")
}
\arguments{
\item{inputname}{the string of the location of the input tree}
}
\value{
data frame containing all information about the input tree file
}
\description{
tree_input_process

\code{tree_input_process} preprocess the phylo object from the input tree file with newick format
}
\examples{

}
