% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_layers_25}
\alias{tree_layers_25}
\title{tree_layers_25}
\format{data frame with 3 columns and 7033 rows
\describe{
\item{x}{x coordinate of the point}
\item{y}{y coordinate of the point}
\item{layer_id}{ id to show which radius the point is on}

...
}}
\usage{
tree_layers_25
}
\description{
A data frame contain information about the tree
created from the sample150.newick file contains taxonomy info of all strains
}
\keyword{datasets}
