% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{treeInputProcess}
\alias{treeInputProcess}
\title{treeInputProcess}
\usage{
treeInputProcess(tree)
}
\arguments{
\item{tree}{the string of the location of the input tree}
}
\value{
data frame containing all information about the input tree file
}
\description{
\code{tree_input_process} preprocess the phylo object from the input tree file with newick format
}
\examples{
treeInputProcess(Tree25)
treeInputProcess(Tree150)

}
