% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{treePlot}
\alias{treePlot}
\title{treePlot

The base tree plotting and a while loop which calling the nodeGroup
and add a nodeGroup each time by adding a layer of the ggplot}
\usage{
treePlot(xy_data, layer_data, npoint)
}
\arguments{
\item{xy_data}{tree data frame with updated x y coordinate information}

\item{layer_data}{layer data frame contaions all layers data}

\item{npoint}{the number of points of each layer}

\item{tip}{this flag is for detecting wjether the point is tip or internal node}
}
\value{
the plot of the tree part
}
\description{
treePlot

The base tree plotting and a while loop which calling the nodeGroup
and add a nodeGroup each time by adding a layer of the ggplot
}
\examples{

}
