% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VF150}
\alias{VF150}
\title{VF150}
\format{A data frame with 150 rows and 4 columns:
\describe{
\item{X}{name reference of the strains}
\item{id}{id of each strain}
\item{vf1}{virulence factor 1 potential of the strain with id1/id2}
\item{vf2}{virulence factor 2 potential of the strain with id1/id2}

...
}}
\usage{
VF150
}
\description{
A dataset contain 2 factors and 150 tree nodes info
}
\keyword{datasets}
