% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VF25_4}
\alias{VF25_4}
\title{VF25_4}
\format{A data frame with 25 rows and 4 columns:
\describe{
\item{X}{name reference of the strains}
\item{id}{id of each strain}
\item{vf1}{virulence factor 1 potential of the strain with id1/id2}
\item{vf2}{virulence factor 2 potential of the strain with id1/id2}
\item{vf3}{virulence factor 3 potential of the strain with id1/id2}
\item{vf4}{virulence factor 4 potential of the strain with id1/id2}

...
}}
\usage{
VF25_4
}
\description{
A dataset contain 4 factors and 25 tree nodes info
}
\keyword{datasets}
