% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xy_df_25}
\alias{xy_df_25}
\title{xy_df_25}
\format{data frame with 11 columns and 49 rows
\describe{
\item{depth}{depth of the node}
\item{id}{ id of this node}
\item{parent}{ the parent of this point / NA if root}
\item{angle}{ angle of this point}
\item{c1}{ child 1 id / -1 if no child}
\item{c2}{ child 2 id / -1 if no child}
\item{c1_a}{ child 1 angle }
\item{c2_a}{ child 2 angle}
\item{depth_length}{ this is the parameter to calculate depth between parent and self if internal node/ length to outer radius if tips }
\item{x}{ x coordinate of the point}
\item{y}{ y coordinate of the point}

...
}}
\usage{
xy_df_25
}
\description{
A data frame contain information about the tree
created from the sample150.newick file contains taxonomy info of all strains
}
\keyword{datasets}
