% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedPlotting.R
\name{combinedPlotting}
\alias{combinedPlotting}
\title{combinePlotting}
\usage{
combinedPlotting(inputCSV, inputTree, inputNum)
}
\arguments{
\item{inputCSV}{input the generated csv file with pathogenic potential information
of each virulence factor of each strain}

\item{inputNum}{the number of input strains}

\item{inputNWK}{input the generated nwk file of the taxonomic information between strains}
}
\value{
Combined the plotting of circular tree and factor ring from the
data of tree(newick file) and table(csv/tsv file)
}
\description{
The main plotting function to combine the two parts of the workflow
}
\details{
This function is used to plot a tree from the newick input, on top of
it plot the factor information according to each strain of the tree
and plot as rings around the tree
}
\examples{

}
