% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{getCoordinates}
\alias{getCoordinates}
\title{getCoordinates}
\usage{
getCoordinates(data1, data2, npoint)
}
\arguments{
\item{data1}{tree data frame without x,y coordinate}

\item{data2}{layer data frame contaions all layers data}

\item{npoint}{the number of points of each layer}
}
\value{
updated tree data frame two new columns of x, y coordinates
}
\description{
This function take input of the tree data and layer data, output the tree
data frame with two more columns contain the x,y coordinate info of each node
}
\examples{

}
