% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{getLayers}
\alias{getLayers}
\title{getLayers

\code{circleFun} call circle function in this function for}
\usage{
getLayers(data, npoint, center = c(0, 0))
}
\arguments{
\item{data}{tree data frame}

\item{npoint}{the number of points of each layer}

\item{center}{the center of the circle}
}
\value{
num_p is the number of points of each layer
}
\description{
getLayers

\code{circleFun} call circle function in this function for
}
\examples{

}
