% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{getNPoints}
\alias{getNPoints}
\title{getNPoints

\code{getNPoints}}
\usage{
getNPoints(data, refine_factor = 100)
}
\arguments{
\item{data}{tree data frame}

\item{refine_factor}{factor of how many points plotted each circle}
}
\value{
num_p is the number of points of each layer
}
\description{
getNPoints

\code{getNPoints}
}
\examples{

}
