% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{getNPoints}
\alias{getNPoints}
\title{getNPoints

This function is used to caculate the number of points for each circular
shaped data frame such as the rings and each circular layer of the tree

The calculation here is based on num_tips is equal to}
\usage{
getNPoints(ntips, refine_factor = 100)
}
\arguments{
\item{ntips}{this is the input size of the strain/tree tips}

\item{refine_factor}{factor of how many points plotted each circle}
}
\value{
The total number of points of each circular shape data
}
\description{
getNPoints

This function is used to caculate the number of points for each circular
shaped data frame such as the rings and each circular layer of the tree

The calculation here is based on num_tips is equal to
}
