% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringPlot.R
\name{getRingData}
\alias{getRingData}
\title{getRingData

This function is used to create a data frame about ring x and y coordinates

\code{circleFun}}
\usage{
getRingData(center_point = c(0, 0), table_data, tree_depth, ring_radius,
  refine_factor_for_ring = 20)
}
\arguments{
\item{center_point}{center of the ring}

\item{table_data}{input the table from table process}

\item{tree_depth}{the max depth of the tree}

\item{ring_radius}{the radius}

\item{refine_factor_for_ring}{refine factor of the plotting ring}
}
\value{
data frame containing x and y coordinate of one ring with npoint index
}
\description{
getRingData

This function is used to create a data frame about ring x and y coordinates

\code{circleFun}
}
\examples{


}
