% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{nodeGroup}
\alias{nodeGroup}
\title{nodeGroup}
\usage{
nodeGroup(tree, layers, idx, plot_arg, npoint, tip = FALSE)
}
\arguments{
\item{tree}{tree data frame with updated x y coordinate information}

\item{layers}{layer data frame contaions all layers data}

\item{idx}{the index of the point in tree data frame}

\item{plot_arg}{pass the base plot to adding more layer on top of it}

\item{npoint}{the number of points of each layer}

\item{tip}{this flag is for detecting whether the point is tip or internal node}
}
\description{
This function is adding a new layer plot of a nodeGroup on the previous plot
each time plot one node. a nodeGroup of the internal node is composed of the node,
the arc/horisontal line along with it and the path to its parent node, the tips
nodeGroup is similar to internal node but without the arc line.
}
