% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{nodeGroup}
\alias{nodeGroup}
\title{nodeGroup

This function is adding node plot of a nodeGroup
each time plot one node. a node group include the node,
the arc around it and the path to its parent node}
\usage{
nodeGroup(data1, data2, idx, plot_arg, npoint, tip = FALSE)
}
\arguments{
\item{data1}{tree data frame with updated x y coordinate information}

\item{data2}{layer data frame contaions all layers data}

\item{idx}{the index of the point in tree data frame}

\item{plot_arg}{pass the base plot to adding more layer on top of it}

\item{npoint}{the number of points of each layer}

\item{tip}{this flag is for detecting wjether the point is tip or internal node}
}
\description{
nodeGroup

This function is adding node plot of a nodeGroup
each time plot one node. a node group include the node,
the arc around it and the path to its parent node
}
\examples{

}
