% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringPlot.R
\name{ringPlot}
\alias{ringPlot}
\title{ringPlot}
\usage{
ringPlot(ring_table_data, radius_data, tree_max_depth,
  present_rate = 0.5, tree_plot)
}
\arguments{
\item{ring_table_data}{data frame from table_processed}

\item{radius_data}{data frame of radius and ring}

\item{tree_max_depth}{the max depth of the tree}

\item{present_rate}{the rate for ploting the factor as present/absent}

\item{tree_plot}{the tree plotting from ggplot}
}
\value{
The ring plot based on the tree plot
}
\description{
This function is used to plot the rings of factoer present/absent of that
specific strain mapped in the tree based on the data from the table,
the same id name mapped to tree data loop the column of the ring_table_data
to plot each ring aroud the tree
}
