% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringPlot.R
\name{ringPlot}
\alias{ringPlot}
\title{ringPlot

This function is used to plot the rings of factoer present/absent of that specific tip
based on the data from the table
loop the column of the table data frame to plot each ring aroud the tree

\code{getRingData}}
\usage{
ringPlot(ring_table_data, radius_data, tree_max_depth,
  present_rate = 0.5, tree_plot)
}
\arguments{
\item{ring_table_data}{data frame from table_processed}

\item{radius_data}{data frame of radius and ring}

\item{tree_max_depth}{the max depth of the tree}

\item{present_rate}{the rate for ploting the factor as present/absent}

\item{tree_plot}{the tree plotting from ggplot}
}
\value{
The ring plot based on the tree plot
}
\description{
ringPlot

This function is used to plot the rings of factoer present/absent of that specific tip
based on the data from the table
loop the column of the table data frame to plot each ring aroud the tree

\code{getRingData}
}
\examples{

}
